requirements.txt
fastapi
uvicorn
pydantic
python-multipart

langchain
langchain-community

pymupdf

faiss-cpu
sentence-transformers